DDP-5603 | added more abstraction for heatmaps
